# ⚡ Improve the parsing performance