--- a/src/types/ParsingContext.ts
+++ b/src/types/ParsingContext.ts
@@ -66,6 +66,5 @@ export function constructContext(
         textDoc: TextDocument.create('dhp://document.mcfunction', 'mcfunction', 0, ''),
         ...custom
     }
-
     return ans
 }