<details>
<summary>Profiling for parsing <code>cpp:tick</code></summary>

```
P [0] 92 ms
P [1] 90 ms
P [2] 85 ms
P [3] 86 ms
P [4] 76 ms
P [5] 86 ms
P [6] 74 ms
P [7] 75 ms
P [8] 86 ms
P [9] 80 ms
P [10] 73 ms
P [11] 93 ms
P [12] 122 ms
P [13] 85 ms
P [14] 85 ms
P [15] 78 ms
P [16] 72 ms
P [17] 82 ms
P [18] 74 ms
P [19] 80 ms
P [20] 89 ms
P [21] 83 ms
P [22] 77 ms
P [23] 90 ms
P [24] 79 ms
P [25] 82 ms
P [26] 84 ms
P [27] 75 ms
P [28] 77 ms
P [29] 76 ms
P [30] 75 ms
P [31] 73 ms
P [32] 84 ms
P [33] 76 ms
P [34] 77 ms
P [35] 71 ms
P [36] 78 ms
P [37] 69 ms
P [38] 74 ms
P [39] 79 ms
P [40] 73 ms
P [41] 74 ms
P [42] 78 ms
P [43] 73 ms
P [44] 73 ms
P [45] 0 ms
P [46] 90 ms
P [47] 77 ms
P [48] 0 ms
P [49] 73 ms
P [50] 77 ms
P [51] 77 ms
P [52] 73 ms
P [53] 71 ms
P [54] 83 ms
P [55] 72 ms
P [56] 0 ms
P [57] 73 ms
P [58] 85 ms
P [59] 75 ms
P [60] 87 ms
P [61] 104 ms
P [62] 73 ms
P [63] 75 ms
P [64] 74 ms
P [65] 78 ms
P [66] 72 ms
P [67] 72 ms
P [68] 87 ms
P [69] 75 ms
P [70] 75 ms
P [71] 82 ms
P [72] 74 ms
P [73] 78 ms
P [74] 85 ms
P [75] 76 ms
P [76] 73 ms
P [77] 78 ms
P [78] 71 ms
P [79] 80 ms
P [80] 76 ms
P [81] 75 ms
P [82] 74 ms
P [83] 76 ms
P [84] 77 ms
P [85] 74 ms
P [86] 73 ms
P [87] 82 ms
P [88] 77 ms
P [89] 74 ms
```

</details>